+++ tests/api/createCar/createCar.spec.js
@@ -70,15 +70,32 @@ test.describe('Car creation API testing', () => {
 
     test.describe('Positive scenario', () => {
 
+        // test.afterEach(async ({ request }) => {
+        //     const carsList = await request.get('/api/cars')
+        //     const { data: cars } = await carsList.json()
+
+        //     await Promise.all(
+        //         cars.map((car) => (async () => {
+        //             const res = await request.delete(`/api/cars/${car.id}`)
+        //             await expect(res).toBeOK()
+        //         })())
+        //     )
+        // })
+
         test.afterEach(async ({ request }) => {
             const carsList = await request.get('/api/cars')
             const { data: cars } = await carsList.json()
 
             await Promise.all(
-                cars.map((car) => (async () => {
-                    const res = await request.delete(`/api/cars/${car.id}`)
-                    await expect(res).toBeOK()
-                })())
+                cars.map(async (car) => {
+                    const checkCar = await request.get(`/api/cars/${car.id}`)
+                    if (checkCar.ok()) {
+                        const res = await request.delete(`/api/cars/${car.id}`)
+                        await expect(res).toBeOK()
+                    } else {
+                        console.log(`Car with ID ${car.id} not found, skipping deletion.`)
+                    }
+                })
             )
         })
